Added last etag storing in database

+++ src/main/kotlin/com/niphoneua/wikipediapreviews/db/DbTables.kt
@@ -7,7 +7,7 @@ object AppInstallation : Table("app_installation") {
     val clientId = varchar("client_id", 36).index(isUnique = true)
     val clientSecret = varchar("client_secret", 64)
     val serverUrl = varchar("server_url", 256)
-    val lastEtag = integer("last_etag").nullable()
+    val lastEtag = long("last_etag").nullable()
 
     override val primaryKey = PrimaryKey(clientId)
 }
